requirements.txt
altair==4.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
boto3==1.14.11
botocore==1.17.11
Brotli==1.0.7
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
dash==1.13.4
dash-core-components==1.10.1
dash-daq==0.5.0
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
dtale==1.8.15
entrypoints==0.3
enum-compat==0.0.3
Flask==1.1.2
Flask-Compress==1.5.0
flask-ngrok==0.0.25
future==0.18.2
giotto-tda==0.2.2
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lz4==3.1.0
MarkupSafe==1.1.1
matplotlib==3.2.2
missingno==0.4.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.5
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.6.0
ppscore==0.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pydeck==0.4.0b2
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-igraph==0.8.2
pytz==2020.1
#pywin32==228
#pywinpty==0.5.7
pyzmq==19.0.1
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.0
Send2Trash==1.5.0
six==1.15.0
streamlit==0.62.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
